refactor: move data fetching to server component and improve component structure
- moved API data fetching to `page.tsx` for better performance
- created `AdvocateTable.tsx` as a client component for filtering
- resolved hydration issues by keeping `page.tsx` as a server component
